LayoutTests/editing/deleting/smart-delete-across-editable-boundaries-2.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This tests for a bug where expansion for smart delete would not consider editable boundaries. Only 'foo' should be deleted. You should see ' bar'. &lt;radr://problem/5390681&gt;</p>
<div contenteditable="true" id="test">foo
<span contenteditable="false"> bar</span></div>

<script src="../editing.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script>

async function editingTest() {
	await UIHelper.selectWordByDoubleTapOrClick(document.getElementById('test'));
	document.execCommand("Delete");
}

runDumpAsTextEditingTest(true);
</script>
</body>
</html>
